requirements.txt
flask
flask-restx
torch
torchvision
filetype
reverse-geocode
exifread
mongoengine
watchdog
opencv-python
numpy
pydantic
pyyaml
